Changed the genCLoseList function to first make sure not to check against the current user, THEN it can begin seeing if users are close enough.
Also, realized that the toRadians() method in the getDistance() function was causing problems. Turns out that's not a real function. I ended up stealing a function from the following Stack Overflow thread:

http://stackoverflow.com/questions/5260423/torad-javascript-function-throwing-error/21623256#21623256

Now the issue is that  my program runs when I click the button on the HTML page, but it doesn't ever get to displaying the data. Debugging time!